+++ Nickel/Framework/Content/ContentManager.cs
@@ -43,17 +43,25 @@ internal sealed class ContentManager
 		this.Enemies = enemies;
 	}
 
-	public static ContentManager Create(Func<ModLoadPhaseState> currentModLoadPhaseProvider, Func<IModManifest, ILogger> loggerProvider, EnumCasePool enumCasePool, IModManifest vanillaModManifest, IModManifest modManagerModManifest, ModDataManager modDataManager)
+	public static ContentManager Create(
+		Func<ModLoadPhaseState> currentModLoadPhaseProvider,
+		Func<IModManifest, ILogger> loggerProvider,
+		ModEventManager eventManager,
+		EnumCasePool enumCasePool,
+		IModManifest vanillaModManifest,
+		IModManifest modLoaderModManifest,
+		ModDataManager modDataManager
+	)
 	{
 		var sprites = new SpriteManager(loggerProvider, enumCasePool, vanillaModManifest);
 		var decks = new DeckManager(currentModLoadPhaseProvider, enumCasePool, vanillaModManifest);
 		var statuses = new StatusManager(currentModLoadPhaseProvider, enumCasePool, vanillaModManifest);
 		var cards = new CardManager(currentModLoadPhaseProvider, loggerProvider, vanillaModManifest);
 		var artifacts = new ArtifactManager(currentModLoadPhaseProvider, loggerProvider, vanillaModManifest);
-		var characters = new CharacterManager(currentModLoadPhaseProvider, loggerProvider, sprites, decks, statuses, vanillaModManifest);
+		var characters = new CharacterManager(currentModLoadPhaseProvider, loggerProvider, eventManager, sprites, decks, statuses, vanillaModManifest, modLoaderModManifest);
 		var parts = new PartManager(enumCasePool, currentModLoadPhaseProvider);
 		var ships = new ShipManager(currentModLoadPhaseProvider, vanillaModManifest);
-		var cardTraits = new CardTraitManager(loggerProvider, vanillaModManifest, modManagerModManifest, modDataManager);
+		var cardTraits = new CardTraitManager(loggerProvider, vanillaModManifest, modLoaderModManifest, modDataManager);
 		var enemies = new EnemyManager(currentModLoadPhaseProvider, loggerProvider, vanillaModManifest);
 		return new(sprites, decks, statuses, cards, artifacts, characters, parts, ships, cardTraits, enemies);
 	}